Fixed bugs #847044 where sphinx-build was hardcoded at /usr/bin/sphinx-build
Read the $PATH environment variable. Split it by : character. For every entry append 
'sphinx-build' and check whether the file exists. If the file exists, check if it 
is executable. If yes, then return the path. If none is found then return None. 
The method calling this search method has to handle None return by showing a WARNING